+++ ui/src/App.tsx
@@ -4,6 +4,8 @@ import type { ContainerInfo, NetworkInfo } from './interfaces/interfaces';
 
 import NetworksPage from './pages/NetworksPage';
 import VisualizerPage from './pages/VisualizerPage';
+// import SankeyPage from './pages/SankeyPage';
+// import ArcPage from './pages/ArcPage';
 import { useAppStore } from './store';
 
 export default function App() {
@@ -117,7 +119,9 @@ export default function App() {
           path='/'
           element={<NetworksPage networks={networks} containers={containers} />}
         />
-        <Route path='/Visualizer'  element={<VisualizerPage networks={networks} containers={containers} />} />
+        <Route path='Visualizer' element={<VisualizerPage networks={networks} containers={containers} />}>
+          
+        </Route>
       </Routes>
     </div>
   );
