+++ ui/src/react-app-env.d.ts
@@ -1,0 +1,4 @@
+declare module "*.png";
+declare module "*.svg";
+declare module "*.jpeg";
+declare module "*.jpg";
